backend/flask/requirements.txt
flask
waitress
flask-cors
ultralytics
opencv-python-headless
pillow
numpy
aiohttp
requests
